requirements.txt
scapy>=2.5.0
rich>=13.0.0
pyyaml>=6.0
pydantic>=2.0.0
watchdog>=3.0.0
pytest>=7.0.0
typer>=0.9.0
fastapi>=0.100.0
uvicorn>=0.23.0
